Refactored controllers and get request for services

+++ Models/Image.cs
@@ -3,13 +3,12 @@ using Postgrest.Models;
 
 namespace WebApi.Models;
 
-// Supabase/Postgrest model mapped to table "images".
 [Table("image")]
 public class Image : BaseModel
 {
     [PrimaryKey("id", false)]
     [Column("id")]
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [Column("name")]
     public string Name { get; set; }
